+++ zcash_client_sqlite/src/wallet.rs
@@ -108,6 +108,8 @@ use zcash_keys::{
         UnifiedIncomingViewingKey, UnifiedSpendingKey,
     },
 };
+#[cfg(zcash_unstable = "nu7")]
+use zcash_primitives::transaction::TxVersion;
 use zcash_primitives::{
     block::BlockHash,
     merkle_tree::read_commitment_tree,
@@ -2890,9 +2892,38 @@ fn parse_tx<P: consensus::Parameters>(
 
         let expiry_height = tx_data.expiry_height();
         if expiry_height > BlockHeight::from(0) {
-            TransactionData::from_parts(
+            let consensus_branch_id = BranchId::for_height(params, expiry_height);
+            #[cfg(zcash_unstable = "nu7")]
+            let tx = if tx_data.version() == TxVersion::V6 {
+                TransactionData::from_parts_v6(
+                    consensus_branch_id,
+                    tx_data.lock_time(),
+                    expiry_height,
+                    tx_data.transparent_bundle().cloned(),
+                    tx_data.sapling_bundle().cloned(),
+                    tx_data.orchard_bundle().cloned(),
+                    tx_data.ironwood_bundle().cloned(),
+                )
+                .freeze()
+            } else {
+                TransactionData::from_parts(
+                    tx_data.version(),
+                    consensus_branch_id,
+                    tx_data.lock_time(),
+                    expiry_height,
+                    #[cfg(all(zcash_unstable = "zfuture", feature = "zip-233"))]
+                    tx_data.zip233_amount(),
+                    tx_data.transparent_bundle().cloned(),
+                    tx_data.sprout_bundle().cloned(),
+                    tx_data.sapling_bundle().cloned(),
+                    tx_data.orchard_bundle().cloned(),
+                )
+                .freeze()
+            };
+            #[cfg(not(zcash_unstable = "nu7"))]
+            let tx = TransactionData::from_parts(
                 tx_data.version(),
-                BranchId::for_height(params, expiry_height),
+                consensus_branch_id,
                 tx_data.lock_time(),
                 expiry_height,
                 #[cfg(all(zcash_unstable = "zfuture", feature = "zip-233"))]
@@ -2902,9 +2933,10 @@ fn parse_tx<P: consensus::Parameters>(
                 tx_data.sapling_bundle().cloned(),
                 tx_data.orchard_bundle().cloned(),
             )
-            .freeze()
-            .map(|t| (expiry_height, t))
-            .map_err(SqliteClientError::from)
+            .freeze();
+
+            tx.map(|t| (expiry_height, t))
+                .map_err(SqliteClientError::from)
         } else {
             Err(SqliteClientError::CorruptedData(
                 "Consensus branch ID not known, cannot parse this transaction until it is mined"